--- a/pumpkin-data/build/fluid.rs
+++ b/pumpkin-data/build/fluid.rs
@@ -670,6 +670,11 @@ pub(crate) fn build() -> TokenStream {
             fn from_props(props: Vec<(String, String)>, fluid: &Fluid) -> Self where Self: Sized;
         }
 
+        pub fn get_fluid(registry_id: &str) -> Option<Fluid> {
+           let key = registry_id.strip_prefix("minecraft:").unwrap_or(registry_id);
+           Fluid::from_registry_key(key)
+        }
+
         impl Fluid {
             #constants