--- a/pumpkin-data/build/fluid.rs
+++ b/pumpkin-data/build/fluid.rs
@@ -594,6 +594,7 @@ pub(crate) fn build() -> TokenStream {
     let properties = property_enums.values().map(|prop| prop.to_token_stream());
 
     quote! {
+        use std::hash::{Hash, Hasher};
         use crate::tag::{Tagable, RegistryKey};
         use pumpkin_util::resource_location::{FromResourceLocation, ResourceLocation, ToResourceLocation};
 
@@ -638,6 +639,20 @@ pub(crate) fn build() -> TokenStream {
             pub flow_distance: u32,
             pub can_convert_to_source: bool,
         }
+
+        impl Hash for Fluid {
+            fn hash<H: Hasher>(&self, state: &mut H) {
+                self.id.hash(state);
+            }
+        }
+
+        impl PartialEq for Fluid {
+            fn eq(&self, other: &Self) -> bool {
+                self.id == other.id
+            }
+        }
+
+        impl Eq for Fluid {}
 
         pub static FLUID_STATES: &[PartialFluidState] = &[
             #(#unique_fluid_states),*