--- a/pumpkin/src/block/blocks/piston/piston_head.rs
+++ b/pumpkin/src/block/blocks/piston/piston_head.rs
@@ -22,7 +22,7 @@ impl PumpkinBlock for PistonHeadBlock {
             .position
             .offset(props.facing.opposite().to_block_direction().to_offset());
         let (new_block, new_state) = args.world.get_block_and_block_state(&pos).await;
-        if PistonBlock::ids(&PistonBlock).contains(&new_block.name) {
+        if &Block::PISTON == new_block || &Block::STICKY_PISTON == new_block {
             let props = PistonProps::from_state_id(new_state.id, new_block);
             if props.extended {
                 // TODO: use player