--- a/pumpkin/src/block/blocks/piston/piston_head.rs
+++ b/pumpkin/src/block/blocks/piston/piston_head.rs
@@ -5,9 +5,9 @@ use pumpkin_macros::pumpkin_block;
 use pumpkin_world::world::BlockFlags;
 
 use crate::block::pumpkin_block::BrokenArgs;
-use crate::block::pumpkin_block::{BlockMetadata, PumpkinBlock};
+use crate::block::pumpkin_block::PumpkinBlock;
 
-use super::piston::{PistonBlock, PistonProps};
+use super::piston::PistonProps;
 
 pub(crate) type PistonHeadProperties = pumpkin_data::block_properties::PistonHeadLikeProperties;