--- a/pumpkin/src/block/registry.rs
+++ b/pumpkin/src/block/registry.rs
@@ -6,6 +6,8 @@ use crate::entity::player::Player;
 use crate::server::Server;
 use crate::world::World;
 use async_trait::async_trait;
+use pumpkin_data::block_properties;
+use pumpkin_data::fluid;
 use pumpkin_data::fluid::Fluid;
 use pumpkin_data::item::Item;
 use pumpkin_data::{Block, BlockDirection, BlockState};
@@ -43,8 +45,8 @@ pub enum BlockActionResult {
 
 #[derive(Default)]
 pub struct BlockRegistry {
-    blocks: HashMap<String, Arc<dyn PumpkinBlock>>,
-    fluids: HashMap<String, Arc<dyn PumpkinFluid>>,
+    blocks: HashMap<u16, Arc<dyn PumpkinBlock>>,
+    fluids: HashMap<u16, Arc<dyn PumpkinFluid>>,
 }
 
 #[async_trait]
@@ -75,7 +77,10 @@ impl BlockRegistry {
         let names = block.names();
         let val = Arc::new(block);
         for i in names {
-            self.blocks.insert(i, val.clone());
+            self.blocks.insert(
+                block_properties::get_block(i.as_str()).unwrap().id,
+                val.clone(),
+            );
         }
     }
 
@@ -83,7 +88,8 @@ impl BlockRegistry {
         let names = fluid.names();
         let val = Arc::new(fluid);
         for i in names {
-            self.fluids.insert(i, val.clone());
+            self.fluids
+                .insert(fluid::get_fluid(i.as_str()).unwrap().id, val.clone());
         }
     }
 
@@ -547,12 +553,12 @@ impl BlockRegistry {
 
     #[must_use]
     pub fn get_pumpkin_block(&self, block: &Block) -> Option<&Arc<dyn PumpkinBlock>> {
-        self.blocks.get(&format!("minecraft:{}", block.name))
+        self.blocks.get(&block.id)
     }
 
     #[must_use]
     pub fn get_pumpkin_fluid(&self, fluid: &Fluid) -> Option<&Arc<dyn PumpkinFluid>> {
-        self.fluids.get(&format!("minecraft:{}", fluid.name))
+        self.fluids.get(&fluid.id)
     }
 
     pub async fn emits_redstone_power(