@topi-banana 2025-07-09
You can use `&'static Block` as the key instead.